Convert all float physics and biological state variables to double precision (64-bit)

+++ Antibiotic.hpp
@@ -2,14 +2,14 @@
 
 class Antibiotic {
 private:
-    float concentration = 0.0f;
+    double concentration = 0.0;
 
 public:
     Antibiotic() = default;
-    explicit Antibiotic(float start_concentration);
+    explicit Antibiotic(double start_concentration);
 
-    float get_concentration() const;
-    void set_concentration(float value);   // новый метод
-    void add(float value);
-    void decrease(float value);
+    double get_concentration() const;
+    void set_concentration(double value);   // новый метод
+    void add(double value);
+    void decrease(double value);
 };
